Renaming get-ir.cpp to get-ir-mwv.cpp
This rename is intended to highlight the fact that the provided program
is a minimal working version: get-ir-mwv.cpp is capable of reading an
SPA file and outputting a CSV file, but the output may be shifted,
incomplete, or may contain errors. The minimal working version is not
intended to work for all SPA files in general, but only those provided
in the SPA-Files folder.