Rewrote source files
Parsing of command-line arguments in main() was entirely redone. This
program now accepts -u, --upper-bound, -l, and --lower-bound as
arguments. These flags only change the data that is written to the CSV
file. Additionally, there is some very simple (a.k.a. unverified) data
processing that computes average % transmittance if a multiple of three
SPA files were passed to the program.